pyproject.toml
[project]
name = "izumi"
version = "0.1.0"
description = "C/C++ embedded source tree OSS detection and SBOM assistance tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cyclonedx-bom>=7.2.2",
    "libclang>=18.1.1",
    "litellm>=1.82.4",
    "pyside6>=6.10.2",
    "spdx-tools>=0.8.5",
]

[project.scripts]
izumi = "main:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-qt>=4.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
